fix:Рекомендации-инт на лонг

+++ src/main/java/ru/yandex/practicum/filmorate/model/Feed.java
@@ -10,7 +10,7 @@ import ru.yandex.practicum.filmorate.util.Operation;
 @NoArgsConstructor
 @EqualsAndHashCode(callSuper = false)
 public class Feed {
-    private int timestamp;
+    private Long timestamp;
     private Integer userId;
     private EventType eventType;
     private Operation operation;
